Limit words to 5, keeping words less than 5
This allows exploring shorter word ladders as well without changing the
limit, but still not taking much of a performance hit by iterating over
all of the longer words. Since 6 letter word ladders are already
rare (according to wikipedia) this is a good breakpoint.

This also fixes an off-by-one reporting error, counting the starting
word as a step in the ladder.